refactor: switch to new JSON Schema everywhere
Fixes MRGFY-4010
Depends-On: https://github.com/Mergifyio/ci-bot/pull/76
