+++ src/components/Tables/ActionOptionsTable.tsx
@@ -1,4 +1,4 @@
-import configSchema from '../../../public/mergify-configuration-schema-future-version.json';
+import configSchema from '../../../public/mergify-configuration-schema.json';
 
 import { OptionDefinition, Def } from './ConfigOptions';
 import { OptionsTableBase } from './OptionsTable';
